flamenco: testing lthash and fixing some off-by-one errors

+++ src/flamenco/snapshot/fd_snapshot_main.c
@@ -321,7 +321,7 @@ do_dump( fd_snapshot_dumper_t *    d,
   /* Create a funk database */
 
   ulong const txn_max =   16UL;  /* we really only need 1 */
-  ulong const rec_max = 1024UL;  /* we evict records as we go */
+  uint const rec_max = 1024UL;  /* we evict records as we go */
 
   ulong funk_tag = 42UL;
   d->funk = fd_funk_join( fd_funk_new( fd_wksp_alloc_laddr( wksp, fd_funk_align(), fd_funk_footprint(txn_max, rec_max), funk_tag ), funk_tag, funk_seed, txn_max, rec_max ) );
